// ENETUM/EWAP000W/src/helpers/interfaces/IAvisosGenerales.ts
export interface ResponseItem {
    [x: string]: any;
    codigoUsuarioModificacion: string;
    descripcionEs: string;
    descripcionEu: string;
    fechaAlta: string;
    fechaBaja: string;
    fechaFinVigencia: string;
    fechaFinVigenciaString: string;
    fechaInicioVigencia: string;
    fechaInicioVigenciaString: string;
    idAviso: number;
    idAvisoExterno: number;
    perfil: string;
    tipoEnlace: string;
    tituloEs: string;
    tituloEu: string;
  }